### 2025-09-19 #9018 @editorialbot
[PRE REVIEW]: GCAMUSAJobs: An R package for employment projections based on GCAM-USA power sector outcomes
**Submitting author:** @yingzhangpnnl (<a href="http://orcid.org/0000-0002-0036-7502">Ying Zhang</a>)
**Repository:** https://github.com/JGCRI/GCAMUSAJobs
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.0.1
