### 2026-02-03 9e36755
update paper requested by JOSS due to scope change
