+++ paper/paper.bib
@@ -1,8 +1,69 @@
+@article{feijoo2020us,
+  title={US energy system transitions under cumulative emissions budgets},
+  author={Feijoo, Felipe and Iyer, Gokul and Binsted, Matthew and Edmonds, James},
+  journal={Climatic Change},
+  volume={162},
+  number={4},
+  pages={1947--1963},
+  year={2020},
+  publisher={Springer}
+}
 
-@software{nrel_jobs_nodate,
+@article{ganji2024implications,
+  title={Implications of climate change mitigation and socioeconomic development on the US electric power sector},
+  author={Ganji, Fatemeh and Liu, Lu and Fellows, Shanna},
+  journal={Energy and Climate Change},
+  volume={5},
+  pages={100125},
+  year={2024},
+  publisher={Elsevier}
+}
+
+@article{ou2018estimating,
+  title={Estimating environmental co-benefits of US low-carbon pathways using an integrated assessment model with state-level resolution},
+  author={Ou, Yang and Shi, Wenjing and Smith, Steven J and Ledna, Catherine M and West, J Jason and Nolte, Christopher G and Loughlin, Daniel H},
+  journal={Applied energy},
+  volume={216},
+  pages={482--493},
+  year={2018},
+  publisher={Elsevier}
+}
+
+@article{pan2025assessment,
+  title={Assessment of mobility decarbonization with carbon tax policies and electric vehicle incentives in the US},
+  author={Pan, Weijie and Shittu, Ekundayo},
+  journal={Applied Energy},
+  volume={379},
+  pages={124838},
+  year={2025},
+  publisher={Elsevier}
+}
+
+@article{zhang2025long,
+  title={Long-term decarbonization impacts on residential energy security across income groups and US states},
+  author={Zhang, Ying and Ou, Yang and Waldhoff, Stephanie and O’Neill, Brian and Iyer, Gokul and Sampedro, Jon and Casper, Kelly},
+  journal={Environmental Research Letters},
+  volume={20},
+  number={5},
+  pages={054040},
+  year={2025},
+  publisher={IOP Publishing}
+}
+
+@Misc{nrel_jobs_nodate,
 	title = {Jobs and Economic Development Impact ({JEDI}) models},
-	url = {https://www.nrel.gov/analysis/jedi/about.html},
-	author = {{NREL}},
+	author = {NLR},
+	year     = {2025},
+	howpublished = {National Laboratory of the Rockies},
+	note      = {\url{https://www.nlr.gov/analysis/jedi/models}}
+}
+
+@Misc{GCAM82,
+  author   = {JGCRI},
+  title    = {{GCAM Documentation (Version 8.2)}},
+  year     = {2025},
+  howpublished = {Pacific Northwest National Laboratory/Joint Global Change Research Institute},
+  note      = {\url{https://github.com/JGCRI/gcam-doc}}
 }
 
 @article{mayfield2023labor,
