+++ paper/paper.md
@@ -36,30 +36,87 @@ affiliations:
 # Summary
 
 The `GCAMUSAJobs` R package was developed to post-process electric power
-projections from GCAM-USA, enabling the estimation of future power sector jobs
+projections from GCAM-USA [@GCAM82], enabling the estimation of future power sector jobs
 at the state-level by generation technology and job type. `GCAMUSAJobs` extends
 GCAM-USA functionality by (1) estimating the capacity levels of different
 activities – operational capacity, capacity addition, and retirement; and (2)
 calculating jobs associated with production activities, including those in
 operation and maintenance (O&M), construction, and decommissioning.
+Additionally, this package is designed to be easily adaptable to the output of
+other energy system models besides GCAM-USA.
 
 # Statement of need
 
 The development of `GCAMUSAJobs` was driven by the need to assess the
-distributional labor impacts of energy system transition
-[@xie2023distributional; @mayfield2023labor; @hanson2023local;
-@raimi2021mapping]. While gross employment [@mayfield2023labor] and power sector
-employment [@xie2023distributional] are expected to grow into the future,  over
-time under both business as usual and decarbonization, @xie2023distributional
-find insignificant differences in power sector jobs between the two scenarios.
-Other research has also suggested that fossil fuel-intensive states may
-experience slower job growth or job losses [@hanson2023local;
+distributional labor impacts of energy system evolution [@xie2023distributional;
+@mayfield2023labor; @hanson2023local; @raimi2021mapping]. While gross employment
+[@mayfield2023labor] and power sector employment [@xie2023distributional] are
+expected to grow into the future, under both business as usual (BAU) and
+alternative scenarios, @xie2023distributional find insignificant differences in
+the U.S. national power sector jobs between BAU and decarbonization scenarios.
+However, distributional differences across U.S. states are much more
+significant, e.g., fossil fuel-intensive states may experience slower job growth
+or job losses, which was also found in other studies [@hanson2023local;
 @mayfield2023labor].
 
-Currently, GCAM-USA does not calculate power sector jobs. `GCAMUSAJobs`
-addresses this gap by providing projected direct power sector jobs based on
-GCAM-USA output, enhancing the functionality of GCAM-USA for labor impact
-analysis.
+The Global Change Analysis Model (GCAM) and its ancillary model GCAM-USA
+[@GCAM82] are powerful tools for studying energy systems dynamics and evolution.
+Many previous studies have applied or integrated the tool to analyze potential
+impacts on the energy system and associated outcomes due to environmental and
+socioeconomic changes (e.g., @feijoo2020us; @ganji2024implications;
+@ou2018estimating; @pan2025assessment; @zhang2025long). Currently, GCAM-USA does
+not calculate power sector jobs. `GCAMUSAJobs` addresses this gap by providing
+projected direct power sector jobs based on GCAM-USA output, enhancing the
+functionality of GCAM-USA for labor impact analysis.
+
+# Software design
+
+The direct purpose of `GCAMUSAJobs` is to calculate different types of jobs in
+the power sector across U.S. states, built upon the existing open-source model
+GCAM-USA. As GCAM-USA is a complex multi-sector model with technological
+details, our design involves in-depth understanding and testing of how various
+components work together in GCAM-USA, especially in the power sector, so that we
+can properly design the jobs calculation in our package and source available
+matching data (e.g., employment factors). This requires irreplaceable human
+efforts to comprehend the existing software and extend its capability through
+collaboration with GCAM-USA modelers and domain experts in socioeconomics.
+
+`GCAMUSAJobs` is also designed to be easily adaptable to the output of broader
+energy system models to enable extended employment analysis. `GCAMUSAJobs`
[…]
