### 2026-02-03 01d5620
paper minor updates
