+++ paper/paper.bib
@@ -54,16 +54,14 @@
 	title = {Jobs and Economic Development Impact ({JEDI}) models},
 	author = {NLR},
 	year     = {2025},
-	howpublished = {National Laboratory of the Rockies},
-	note      = {\url{https://www.nlr.gov/analysis/jedi/models}}
+	howpublished = {National Laboratory of the Rockies. Available at: \url{https://www.nlr.gov/analysis/jedi/models}}
 }
 
 @Misc{GCAM82,
   author   = {JGCRI},
   title    = {{GCAM Documentation (Version 8.2)}},
   year     = {2025},
-  howpublished = {Pacific Northwest National Laboratory/Joint Global Change Research Institute},
-  note      = {\url{https://github.com/JGCRI/gcam-doc}}
+  howpublished = {Joint Global Change Research Institute, Pacific Northwest National Laboratory. Available at: \url{https://github.com/JGCRI/gcam-doc}}
 }
 
 @article{mayfield2023labor,
