### 2026-02-17 #9018 @ibarraespinosa
Hei have no time at this time

Thanks

On Tue, Feb 17, 2026, 10:35 AM Fangzhou Xie ***@***.***>
wrote:

> *fangzhou-xie* left a comment (openjournals/joss-reviews#9018)
> <https://github.com/openjournals/joss-reviews/issues/9018#issuecomment-3916080088>
>
> Hi @RBhupi <https://github.com/RBhupi>, would you be willing to review
> this submission to JOSS? We carry out our checklist-driven reviews here in
> GitHub issues and follow these guidelines
> <https://joss.readthedocs.io/en/latest/review_criteria.html>.
>
