### 2026-02-26 #9018 @MatthewHeun
> [@MatthewHeun](https://github.com/MatthewHeun), would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow [these guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html).

Unfortunately, I am not in a position to perform this review at this time. I have a high teaching load this semester. After May, I am much more available. Sorry!
