### 2026-03-04 #9018 @rowlandseymour
> [@rowlandseymour](https://github.com/rowlandseymour), would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow [these guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html).

Yes, I can. Looks interesting.
