### 2026-03-04 #10142 @editorialbot
[REVIEW]: GCAMUSAJobs: An R package for employment projections based on GCAM-USA power sector outcomes
**Submitting author:** @yingzhangpnnl (<a href="http://orcid.org/0000-0002-0036-7502">Ying Zhang</a>)
**Repository:** https://github.com/JGCRI/GCAMUSAJobs
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.1.1
