### 2026-03-04 #10142 @fangzhou-xie
👋 @klau506 @rowlandseymour - thanks for agreeing to review, and this is the thread for reviewing the article. While the author(s) @yingzhangpnnl are working on the minor revision on the paper, I believe we could start the review process for the software.

The JOSS review is different from most other journals. Our goal is to work with the authors to help them meet our criteria instead of merely passing judgment on the submission. As such, the reviewers are encouraged to submit issues and pull requests on the software repository. When doing so, please mention #10142 so that a link is created to this thread (and I can keep an eye on what is happening). Please also feel free to comment and ask questions on this thread. In my experience, it is better to post comments/questions/suggestions as you come across them instead of waiting until you've reviewed the entire package.
