### 2026-03-05 a769e60
add 'state of the field' and missing DOIs
