+++ paper/paper.bib
@@ -6,7 +6,8 @@
   number={4},
   pages={1947--1963},
   year={2020},
-  publisher={Springer}
+  publisher={Springer},
+  DOI={10.1007/s10584-020-02670-0}
 }
 
 @article{ganji2024implications,
@@ -16,7 +17,8 @@
   volume={5},
   pages={100125},
   year={2024},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  DOI={10.1016/j.egycc.2024.100125}
 }
 
 @article{ou2018estimating,
@@ -26,7 +28,8 @@
   volume={216},
   pages={482--493},
   year={2018},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  DOI={10.1016/j.apenergy.2018.02.122}
 }
 
 @article{pan2025assessment,
@@ -36,7 +39,8 @@
   volume={379},
   pages={124838},
   year={2025},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  DOI={10.1016/j.apenergy.2024.124838}
 }
 
 @article{zhang2025long,
@@ -47,7 +51,8 @@
   number={5},
   pages={054040},
   year={2025},
-  publisher={IOP Publishing}
+  publisher={IOP Publishing},
+  DOI={10.1088/1748-9326/adca4c}
 }
 
 @Misc{nrel_jobs_nodate,
@@ -71,7 +76,8 @@
   volume={177},
   pages={113516},
   year={2023},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  DOI={10.1016/j.enpol.2023.113516}
 }
 
 @article{rutovitz2015calculating,
@@ -89,7 +95,8 @@
   number={4},
   pages={044026},
   year={2020},
-  publisher={IOP Publishing}
+  publisher={IOP Publishing},
+  DOI={10.1088/1748-9326/ab506d}
 }
 
 @techreport{Kennedy2024renewable,
@@ -103,7 +110,8 @@
   title={Local labor market impacts of the energy transition: prospects and policies},
   author={Hanson, Gordon H},
   year={2023},
-  institution={National Bureau of Economic Research}
+  institution={National Bureau of Economic Research},
+  DOI={10.3386/w30871}
 }
 
 @article{raimi2021mapping,
@@ -122,7 +130,8 @@
   number={1},
   pages={108--121},
   year={2017},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  DOI={10.1016/j.joule.2017.07.005}
 }
 
 @article{xie2023distributional,
@@ -133,7 +142,8 @@
   number={11},
   pages={1203--1212},
   year={2023},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  DOI={10.1038/s41558-023-01802-5}
 }
 
 @article{edmonds2017global,
@@ -202,7 +212,8 @@
   volume={6},
   number={PNNL-SA-131044},
   year={2018},
-  publisher={Pacific Northwest National Lab.(PNNL), Richland, WA (United States)}
+  publisher={Pacific Northwest National Lab.(PNNL), Richland, WA (United States)},
+  DOI={10.5334/jors.208}
 }
 @article{carrol1988,
   title={ Transformation and Weighting in Regression},
@@ -277,5 +288,6 @@
   number={3},
   pages={325--332},
   year={1991},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  DOI={10.1093/imanum/11.3.325}
 }
