### 2026-03-05 dsheng1026/GOFREE#2 @rowlandseymour
Multiple cre tags
https://github.com/JGCRI/GCAMUSAJobs/blob/a769e6016abc84f49b51f410202f1100d8e53f34/DESCRIPTION#L4

I'm having issues installing this package. I think it's because there are two creators listed, but packages are only allowed one. When running this command `devtools::install_github("JGCRI/GCAMUSAJobs")`, I get the following error:

> Error : Invalid DESCRIPTION file
>
> Authors@R field gives more than one person with maintainer role:
>   Di Sheng <di.sheng@pnnl.gov> [aut, cre]
>   Ying Zhang <ying.zhang@pnnl.gov> [cre] (<https://orcid.org/0000-0002-0036-7502>)
>
> See section 'The DESCRIPTION file' in the 'Writing R Extensions'
> manual.
> ERROR: installing package DESCRIPTION failed for package ‘GCAMUSAJobs’
> * removing ‘/Library/Frameworks/R.framework/Versions/4.4-arm64/Resources/library/GCAMUSAJobs’

https://github.com/openjournals/joss-reviews/issues/10142#issuecomment-4007306799
