### 2026-03-05 dsheng1026/GOFREE#3 @rowlandseymour
Contributing file
One of the JOSS requirements is a file outlining how to contribute. For example, you can find the ggplot2 CONTRIBUTING file here: https://ggplot2.tidyverse.org/CONTRIBUTING.html

https://github.com/openjournals/joss-reviews/issues/10142#issuecomment-4007306799
