### 2026-03-05 dsheng1026/GOFREE#4 @rowlandseymour
Cut instead of if else
I would say this is optional, as it is more style than anything else, but would it be better to use cut here that lots of if else statements? It reduces redundancy in the if else statements, is easier to maintain and outputs a factor directly. If I am correct, something like this should work:

```
breaks <- c(0, 1, 5, 10, 25, 50, 75, 100, 200, 300, Inf)
labels <- c("(0, 1k]", "(1k, 5k]", "(5k, 10k]", "(10k, 25k]",
            "(25k, 50k]", "(50k, 75k]", "(75k, 100k]",
            "(100k, 200k]", "(200k, 300k]", "> 300k")

df.map$bin <- cut(df.map$value, breaks = breaks, labels = labels)
```

https://github.com/openjournals/joss-reviews/issues/10142#issuecomment-4007306799
