### 2026-03-05 dsheng1026/GOFREE#5 @rowlandseymour
Improve rigour of test
https://github.com/JGCRI/GCAMUSAJobs/blob/014e998e83fe08262f1b546570bf5855764dc3c5/R/plots.R#L261

There are some tests here, which is good, but they could be a bit more rigourours. For example, these ones just check the column headings are correct (for both EJ and GW). Would it be worth adding a couple more tests, for example checking that the units in _GW and GW, and that the _EJ and _GW function are different?

https://github.com/openjournals/joss-reviews/issues/10142#issuecomment-4007306799
