### 2026-04-07 dsheng1026/GOFREE#6 @klau506
Manuscript suggestions
As part of the [JOSS review process](https://github.com/openjournals/joss-reviews/issues/10142#issuecomment-4007307303), find below some suggestions to improve the manuscript:

### SUMMARY
- L9: Since it's the Summary section, maybe remove all references. Although I leave this up to the Editor.
- L11 & L12: Will suggest to remove the numbers. Without them the sentence is easy to read, while their addition give to the Summary a more technical aspect.
- L14: Consider also mentioning "integrated assessment models" in addition to "energy models"

### STATEMENT OF NEED
- As a genaral comment on this section: I would recomment restructuring it by 1st, detailing the importance and the expected evolution of employment and power sector, with the differences across scenarios and geographic areas (similar to the current first paragraph but without the first sentence); 2nd, briefly introducing GCAM-USA; 3rd, indicating that this IAM does not account for the power sector jobs (also maybe indicate if there are other IAMs that do so), and thus, the gap that GCAMUSAJobs will bridge.
- L19: Very long sentence and repetitve: what's between the commas " under both business as usual (BAU) and alternative scenarios, " it's the same that comes after "@xie2023distributional find insignificant differences in the U.S. national power sector jobs between BAU and decarbonization scenarios". Consider rephrasing and simplifying
- L21: Spell the accronym at its first apperance (U.S). Notice also that later you refer to GCAM-USA, so maybe consider "USA" instead of "U.S." for consistency
- L22: Change "However" to "Moreover", since it is an addition of differences in the sector jobs (first scenario differences, second geographical differences)
- L27: Change "integrated multi-sector models" to "integrated assessment models". This is the technical term for GCAM, an IAM, right? Considering the right wording will ensure other researchers could find easily your paper and use the package

### STATE OF THE FIELD
- Could be nice to have some background about Jobs in other IAMs. In this way, you could split the section into (a) within IAMs Jobs module and (b) post-processing Jobs tools
- L36: Consider "IAM" instead of "integrated multi-sector model"

### SOFTWARE DESIGN
- As a general comment: the first paragraph is not very informative for the reader/user. It basically details the multidiciplinary team required to develop the package, but not the interal software design of the tool. Consider removing it.
  In the second paragraph, you detail at the beginning how to extend its functionality with other energy models. First, would be nice to know how the basics of the package. Consider restructuring this section including the Workflow and Key functions sections that you created later.  In the same line, would be good to include a reference to the Methods vignette, which is very well written and detailed.
- L47: Consider specifying if the "different types of jobs" are estimated as number of workers, economic benefit, number of companies...
- L48: Again, U.S. acronym (maybe USA?). Throughout the whole paper (e.g., L69...)

### RESEARCH IMPACT STATEMENT
- L69: Sources of impact include (remove "can")
- Besides indicating past usage, indicate also future usage expectations

### WORKFLOW
- Fig 1: Typo in the caption (double "Figure 1" text)
- It is not clear to me which are exactly the GCAM-USA outputs that are used as inputs of the R pkg. The colors are misleading, since green inputs are leading to other inputs. Could be nice to redo the paragraph ennumerating the inputs from GCAM-USA (if more than one), the data to deal with the estimations and internal computations (what you call assumptions), and the outputs. Also, for the output "Job", indicate the units.

### KEY FUNCTIONS
- L111: Add link to the further documentation of this list
- L113: Add link to the mentioned vignette
- This section is very nicely written and very informative about the interla package structure. Could be great to use it as Software Design section

### IMPLEMENTATION
- Fig 2: Typo in the caption (double "Figure 2" text)
- This section could be integrated into Research Impact Statement section. You could redo the section by indicating (a) academic analysis performed with the pkg (so, the Implementation example); (b) the real world-usage example (in the Research Impact section, the UniMaryland usage); and (c), expected future usage
