### 2026-04-07 dsheng1026/GOFREE#7 @klau506
Code suggestions
As part of the [JOSS review process](https://github.com/openjournals/joss-reviews/issues/10142#issuecomment-4007307303), find below some suggestions to improve the code:

- In general, there are "TODO" comments in the scripts. Consider removing them.
- Consider further detailing the inputs of the functions. For example, in `R/ReadGCAM.R` you have comments about the input path in L23 which could be helpful in L5.
- The `R/constants.R` script, could be moved to `inst/extdata/saveDataFiles.R` file, using the usethis package and the `use_data` function. See for instance [this package](https://github.com/bc3LC/gcamreport/blob/gcam-core/inst/extdata/saveDataFiles_GCAM7.0.R)
- The subfolder `inst/extdata/GCAM/GCAM6` contains the same files that `inst/extdata/GCAM`. Consider avoiding duplicates
