### 2026-04-07 dsheng1026/GOFREE#8 @klau506
Documentation suggestions
As part of the [JOSS review process](https://github.com/openjournals/joss-reviews/issues/10142#issuecomment-4007307303), find below some suggestions to improve the manuscript:

## README
- The Content links do not work (at least the Installation Guide link)
- Add a link to the provided step by step tutorial on the Documentation
- Add a link to the "how to contribute" section on the Documentation
- The installation using devtools crashes
```
─  preparing 'GCAMUSAJobs': (1.9s)
E  checking DESCRIPTION meta-information ...
   Authors@R field gives more than one person with maintainer role:
     Di Sheng <di.sheng@pnnl.gov> [aut, cre]
     Ying Zhang <ying.zhang@pnnl.gov> [cre] (<https://orcid.org/0000-0002-0036-7502>)

   See section 'The DESCRIPTION file' in the 'Writing R Extensions'
   manual.
Error: Failed to install 'GCAMUSAJobs' from GitHub:
  ! System command 'Rcmd.exe' failed
```

The same error occurs when cloning the repo and using `devtools::install_all()`. Using R4.4.0 version.

## WEBSITE
- The info in the documentation website is great! However, it is difficult to navigate through the different sections. Consider reorganising/renaming the Articles Tab and subtabs (e.g., package_vignette could be names as Tutorial or Step by step example). As an idea, but any modifications will be nice: the Limitations could be a subsection of Methods; the User Notice could be a different tab; and you could create a new Tutorials/Usage tab, with the Examples (currently package_vignette) and the Case Study?

#### How to Contribute section
Currently missing. Indicate in the Doc page guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

#### Methods
- Consider using scientific style in the text. For e.g., ... installed capacity in period t, .... , write the "t" between dollar symbols "$t$".
- Consider adding some bold words to facilitate the reader go through the text.
- Capacity calculation: "... generation by activity (a)". This "(a)" does not appear anywhere else, what are you refearing to? Consider adding the abbrevations in the subtitles, e.g., "op" in the Power from operational capacity, "in" in the Potential generation from non-retired installed capactiy, and so on.
- Typo in the last line {EF}_{2050}

#### Tutorial
Will love to test it once I can install the pkg ;)

## Tests
Currently missing. Consider a build test and a code coverage test
