### 2026-04-07 #10142 @klau506
Hi @fangzhou-xie, @yingzhangpnnl, and @dsheng1026,

First of all, please accept my apologies for the big delay in the revision process. I would also like to thank you for the opportunity to review this impressive tool; it’s been a pleasure exploring it!!

I have completed my review, focusing on the Manuscript (see the open issue [here](https://github.com/JGCRI/GCAMUSAJobs/issues/6)), the Documentation (see the open issue [here](https://github.com/JGCRI/GCAMUSAJobs/issues/8)), and the Code (see the open issue [here](https://github.com/JGCRI/GCAMUSAJobs/issues/7)). As noted in the Documentation issue, I encountered some difficulties during installation, which meant I wasn't able to run the tutorial. I would be very happy to test it again once that is resolved.

Please let me know if you have any questions or if you would like me to conduct a second round of review.

Best regards,
Klàu
