+++ DESCRIPTION
@@ -2,12 +2,12 @@ Package: GCAMUSAJobs
 Title: Tool for project jobs based on GCAM-USA results
 Version: 0.0.1
 Authors@R: c( 
-    person("Di", "Sheng", , "di.sheng@pnnl.gov", role = c("aut", "cre")),
-    person("Ying", "Zhang", , "ying.zhang@pnnl.gov", role = "cre",
+    person("Di", "Sheng", , "di.sheng@pnnl.gov", role = c("aut")),
+    person("Ying", "Zhang", , "ying.zhang@pnnl.gov", role = c("aut","cre"),
            comment = c(ORCID = "0000-0002-0036-7502"))
            )
 Description: This packge projects state-level power sector employment by fuel type and occupation, based on GCAM-USA power sector output.
-License: BSD 2-Clause License + file LICENSE
+License: BSD_2_clause + file LICENSE
 URL: https://jgcri.github.io/GCAMUSAJobs
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
@@ -19,7 +19,7 @@ Imports:
     rgcam,
     tidyr,
     ggplot2,
-    sf
+    tibble
 Depends: 
     R (>= 2.10)
 LazyData: true
