### 2026-06-11 6ef28c0
update R scripts
to pass devtools::check(); clean notes; remove repeated lines and add output variable (now in list format) for downstream uses
