+++ inst/extdata/GCAM/Assumption.R
@@ -20,10 +20,8 @@ GCAM_model_step <- 5
 # Data source for hydropower capital cost
 # https://hub.globalccsinstitute.com/publications/renewable-power-generation-costs-2012-overview/51-hydropower-capital-costs
 
-#TODO: hydro and geothermal capacity factor for USA or USA states
-
 # used in plutus
-# hydro_cap_fact <- 0.38 
+# hydro_cap_fact <- 0.38
 
 # https://www.irena.org/-/media/Files/IRENA/Agency/Publication/2012/RE_Technologies_Cost_Analysis-HYDROPOWER.pdf
 # P16. range from 23% to 95%, average is 0.5
@@ -31,41 +29,40 @@ GCAM_model_step <- 5
 # large hydro Cap.Fac.(%) ranges from 25 to 90
 # small hydro Cap.Fac.(%) ranges from 20 to 95
 
-# P16 
+# P16
 # One option: a high installed capacity and low capacity factor to provide electricity predominantly, nonbase
 # Alternate: a low installed capacity and high capacity factors less flexibility, base
 
-# hydro_cf <- 0.9 # upper 
+# hydro_cf <- 0.9 # upper
 # hydro_cf <- 0.5 # average
 
 # https://www.ipcc.ch/site/assets/uploads/2018/03/Chapter-5-Hydropower-1.pdf
-# P10 
+# P10
 # capacity factor average for North_Am 47%
 # hydro_cf <- 0.47
 
-# Hall et al., 2003 
+# Hall et al., 2003
 # Estimation of economic parameters of U.S. hydropower resources
 # https://www.osti.gov/servlets/purl/1218138
 # USA ranges from 41%-61%
 # Table 8 has the state level monthly capacity factor (exported as csv, see Assumptions/hydro_CF_USA.csv)
 
-hydro_cf <- read.csv("Assumptions/hydro_CF_USA.csv", skip = 3, header = T) %>% 
-  select(region = Abbr, value = Annual) %>% 
+hydro_cf <- read.csv("Assumptions/hydro_CF_USA.csv", skip = 3, header = T) %>%
+  select(region = Abbr, value = Annual) %>%
   mutate(fuel = "hydro")
 
 # https://www.eia.gov/todayinenergy/detail.php?id=42036
 # "Geothermal capacity factors in the United States averaged 76% in 2018"
-geo_cf <- 0.76 
+geo_cf <- 0.76
 
 # another source of hydrogen, solar and wind capacity factor in the U.S.
 #https://energy.mit.edu/wp-content/uploads/2023/04/MITEI-WP-2023-02.pdf
 
 
 # Decommission ----
-#TODO: gather non JEDI job data into a table
 # currently use construction EF (JEDI based) as decom EF
 # decomm.EF <- 1.65 # job-yrs/MW
-# use the construction EF from JEDI 
+# use the construction EF from JEDI
 
 # backup electricity capacity factor
 backup_cf <- 0.05 # assumption in GCAM
@@ -77,25 +74,25 @@ read.GCAM.csv
 A23.globaltech_retirement <- read.GCAM.csv("A23.globaltech_retirement")
 A23.elecS_tech_mapping_cool <- read.GCAM.csv("A23.elecS_tech_mapping_cool")
 
-L2233.StubTechSCurve_elecS_cool_USA <- read.GCAM.csv("L2233.StubTechSCurve_elecS_cool_USA") %>% 
-  dplyr::select(-X) %>% 
+L2233.StubTechSCurve_elecS_cool_USA <- read.GCAM.csv("L2233.StubTechSCurve_elecS_cool_USA") %>%
+  dplyr::select(-X) %>%
[…]
