### 2026-06-11 d57049e
update DESCRIPTION and LICENSE
