+++ paper/paper.md
@@ -36,60 +36,67 @@ affiliations:
 # Summary
 
 The `GCAMUSAJobs` R package was developed to post-process electric power
-projections from GCAM-USA [@GCAM82], enabling the estimation of future power sector jobs
+projections from GCAM-USA, enabling the estimation of future power sector jobs
 at the state-level by generation technology and job type. `GCAMUSAJobs` extends
-GCAM-USA functionality by (1) estimating the capacity levels of different
-activities – operational capacity, capacity addition, and retirement; and (2)
-calculating jobs associated with production activities, including those in
-operation and maintenance (O&M), construction, and decommissioning.
-Additionally, this package is designed to be easily adaptable to the output of
-other energy system models besides GCAM-USA.
+GCAM-USA functionality by estimating the capacity levels of different activities
+– operational capacity, capacity addition, and retirement; and calculating jobs
+associated with those activities, including operation and maintenance (O&M),
+construction, and decommissioning. Additionally, this package is designed to be
+easily adaptable to the output of other energy system models and Integrated
+Assessment Models (IAMs) besides GCAM-USA.
 
 # Statement of need
 
-The development of `GCAMUSAJobs` was driven by the need to assess the
-distributional labor impacts of energy system evolution [@xie2023distributional;
-@mayfield2023labor; @hanson2023local; @raimi2021mapping]. While gross employment
-[@mayfield2023labor] and power sector employment [@xie2023distributional] are
-expected to grow into the future, under both business as usual (BAU) and
-alternative scenarios, @xie2023distributional find insignificant differences in
-the U.S. national power sector jobs between BAU and decarbonization scenarios.
-However, distributional differences across U.S. states are much more
-significant, e.g., fossil fuel-intensive states may experience slower job growth
-or job losses, which was also found in other studies [@hanson2023local;
-@mayfield2023labor].
+It is important to assess the distributional labor impacts of energy system
+evolution, which can affect labor demand across regions, technologies, and skill
+groups [@xie2023distributional; @mayfield2023labor; @hanson2023local;
+@raimi2021mapping]. While gross employment [@mayfield2023labor] and power sector
+employment [@xie2023distributional] are expected to grow into the future under
+both business as usual (BAU) and alternative scenarios, these aggregate trends
+mask important distributional differences. @xie2023distributional find
+insignificant differences in the United States (U.S.) national power sector jobs
+between BAU and decarbonization scenarios; however, distributional differences
+across U.S. states are much more significant. For example, fossil fuel-intensive
+states may experience slower job growth or job losses, which was also found in
+other studies [@hanson2023local; @mayfield2023labor].
 
-The Global Change Analysis Model (GCAM) and its ancillary model GCAM-USA
-[@GCAM82] are integrated multi-sector models that focus on the interaction among
-human activities, technology, and physical systems. They are powerful tools for
-studying energy systems dynamics and evolution. Many previous studies have
-applied the model to analyze potential impacts on the energy system and
-associated outcomes due to environmental and socioeconomic changes [e.g.,
-@feijoo2020us; @ganji2024implications; @ou2018estimating; @pan2025assessment;
-@zhang2025long]. Currently, GCAM-USA does not calculate power sector jobs.
-`GCAMUSAJobs` addresses this gap by providing projected direct power sector jobs
-based on GCAM-USA output, enhancing the functionality of GCAM-USA for labor
-impact analysis.
+The Global Change Analysis Model (GCAM) and its regional extension GCAM-USA with
+state-level details [@GCAM82] are integrated multi-sector models that focus on
+the interaction among human activities, technology, and physical systems. They
+are powerful tools for studying energy systems dynamics and evolution. Many
[…]
