+++ paper/paper.bib
@@ -291,3 +291,36 @@
   publisher={Oxford University Press},
   DOI={10.1093/imanum/11.3.325}
 }
+@article{pai_meeting_2021,
+	title = {Meeting well-below 2°{C} target would increase energy sector jobs globally},
+	volume = {4},
+	issn = {2590-3322},
+	DOI = {10.1016/j.oneear.2021.06.005},
+	abstract = {Summary To limit global warming to well-below 2°C (WB2C), fossil fuels must be replaced by low-carbon energy sources. Support for this transition is often dampened by the impact on fossil fuel jobs. Previous work shows that pro-climate polices could increase employment by 20 million net energy jobs, but these studies rely on Organisation for Economic Co-operation and Development (OECD) jobs data, assumptions about jobs in non-OECD countries, and a single baseline assumption. Here we combine a global dataset of job intensities across 11 energy technologies and five job categories in 50 countries with an integrated assessment model under three shared socioeconomic pathways. We estimate direct energy jobs under a WB2C scenario and current policy scenarios. We find that, by 2050, energy sector jobs would grow from today's 18 million to 26 million under a WB2C scenario compared with 21 million under the current policy scenario. Fossil fuel extraction jobs would rapidly decline, but losses will be compensated by gains in solar and wind jobs, particularly in the manufacturing sector (totaling 7.7 million in 2050).},
+	number = {7},
+	journal = {One Earth},
+	author = {Pai, Sandeep and Emmerling, Johannes and Drouet, Laurent and Zerriffi, Hisham and Jewell, Jessica},
+	year = {2021},
+	pages = {1026--1036},
+}
+@article{malik_climate_2021,
+	title = {Climate policy accelerates structural changes in energy employment},
+	volume = {159},
+	issn = {0301-4215},
+	DOI = {10.1016/j.enpol.2021.112642},
+	journal = {Energy Policy},
+	author = {Malik, Aman and Bertram, Christoph and Kriegler, Elmar and Luderer, Gunnar},
+	year = {2021},
+	pages = {112642},
+}
+@article{emmerling_green_2025,
+	title = {Green jobs and just transition: {Employment} implications of {Europe}'s {Net} {Zero} pathway},
+	volume = {127},
+	issn = {2214-6296},
+	DOI = {10.1016/j.erss.2025.104292},
+	journal = {Energy Research \& Social Science},
+	author = {Emmerling, Johannes and Drouet, Laurent and Jewell, Jessica and Nacke, Lola and Pai, Sandeep and Zerriffi, Hisham},
+	year = {2025},
+	pages = {104292},
+}
+
