### 2026-06-11 dsheng1026/GOFREE#6 @klau506
Thank you for taking care of all these issues. A last minor comment for the Manuscript:
L115: you refer to GCAM-USA instead of GCAM, right?
