### 2026-06-11 dsheng1026/GOFREE#8 @klau506
Thank you for taking care of all these issues! The installation and tutorial worked smoothly :)

A minor comment regarding the Step by step tutorial. In [Step 8](https://jgcri.github.io/GCAMUSAJobs/articles/package_vignette.html#step8-plot-job-by-job-types), consider indicating "state level results (e.g., Colorado)" instead of simply "Colorado" to facilitate comprehension.
