### 2026-06-11 dsheng1026/GOFREE#7 @klau506
Thank you so much for the effort of structuring and tidying the code. A small comment: I assume that to use the `inst/extdata/GCAM/GCAM6` files, you need to run `Assumptions.R` setting the directory to that folder (to distinguish from the data stored in the `inst/extdata/GCAM` folder. If not indicated in the documentation (which I could not find), would be nice to indicate it in a short paragraph at the beggining of the `Assumptions.R` script.
