### 2026-06-11 dsheng1026/GOFREE#7 @klau506
Besides, I find that github actions automated tests are missing. Could you consider including some automated testing?
