### 2026-06-11 dsheng1026/GOFREE#4 @yingzhangpnnl
Thanks for the suggestion! We kept it as ifelse because this function is used widely in GCAM-USA. So users of this package should be familiar with it too.
