### 2026-06-11 dsheng1026/GOFREE#6 @yingzhangpnnl
Yes, that's right! GCAM and GCAM-USA are released together so GCAM v7.1 package contains GCAM-USA in it. I've updated it to GCAM-USA v7.1 to be clearer. Thanks!
