### 2026-06-11 dsheng1026/GOFREE#8 @yingzhangpnnl
Thanks! Agree it's a little confusing. I've updated it to "State level (e.g., Colorado)" to be consistent in format with the "National level" above it. I've updated them in Step 8, as well as in Step 6&7.
