### 2026-06-12 48f74fe
update ReadGCAM and add notes in Assumptions.R
