### 2026-06-12 c034a9c
Create R-CMD-check.yaml
