+++ .github/workflows/R-CMD-check.yaml
@@ -15,13 +15,16 @@ jobs:
 
     steps:
       - uses: actions/checkout@v4
+
       - uses: r-lib/actions/setup-r@v2
         with:
           use-public-rspm: true
+
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
           extra-packages: |
             any::rcmdcheck
             any::remotes
-        needs: check
+          needs: check
+
       - uses: r-lib/actions/check-r-package@v2
