+++ man/years_mapping.Rd
@@ -1,0 +1,16 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/data.R
+\docType{data}
+\name{years_mapping}
+\alias{years_mapping}
+\title{GCAM year mapping data}
+\format{
+A data frame.
+}
+\usage{
+years_mapping
+}
+\description{
+Mapping between GCAM model years and related time periods.
+}
+\keyword{datasets}
