+++ man/CONV_EJ_KWH.Rd
@@ -1,0 +1,16 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/data.R
+\docType{data}
+\name{CONV_EJ_KWH}
+\alias{CONV_EJ_KWH}
+\title{Conversion factor from exajoules to kilowatt-hours}
+\format{
+A numeric value.
+}
+\usage{
+CONV_EJ_KWH
+}
+\description{
+A numeric conversion factor used to convert energy values from EJ to kWh.
+}
+\keyword{datasets}
