+++ man/convMW_kW.Rd
@@ -1,0 +1,16 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/data.R
+\docType{data}
+\name{convMW_kW}
+\alias{convMW_kW}
+\title{Conversion factor from megawatts to kilowatts}
+\format{
+A numeric value.
+}
+\usage{
+convMW_kW
+}
+\description{
+A numeric conversion factor used to convert MW to kW.
+}
+\keyword{datasets}
