+++ man/convGW_MW.Rd
@@ -1,0 +1,16 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/data.R
+\docType{data}
+\name{convGW_MW}
+\alias{convGW_MW}
+\title{Conversion factor from gigawatts to megawatts}
+\format{
+A numeric value.
+}
+\usage{
+convGW_MW
+}
+\description{
+A numeric conversion factor used to convert GW to MW.
+}
+\keyword{datasets}
