+++ man/trial_data.Rd
@@ -1,0 +1,16 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/data.R
+\docType{data}
+\name{trial_data}
+\alias{trial_data}
+\title{Example trial data}
+\format{
+A data frame.
+}
+\usage{
+trial_data
+}
+\description{
+Example data used for testing or demonstration.
+}
+\keyword{datasets}
