### 2026-06-12 dsheng1026/GOFREE#7 @yingzhangpnnl
Thanks! We've added an "Actions" workflow using R CMD check to automatically run package checks and the existing "testthat" tests on pushes and pull requests to the main branch. We also addressed the documentation warnings identified by the automated check so that the workflow can complete successfully.
