### 2026-06-12 dsheng1026/GOFREE#7 @yingzhangpnnl
I forgot to mention that we've also added comments in the Assumptions.R as suggested.
