### 2026-06-15 3224370
fix typo in DESCRIPTION
