### 2026-08-18 #10142 @samhforbes
Hi @yingzhangpnnl can I check - I think R packages will normally require the appropriate LICENSE.md as well as the LICENSE file. See [https://r-pkgs.org/license.html](https://r-pkgs.org/license.html). I can only see the one for you (apologies if I'm missing something) - it should match the one in the zenodo repo.
